requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
astor==0.8.1
async-timeout==4.0.2
autopep8==2.0.1
blobfile==2.0.1
certifi==2023.5.7
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.1
cmake==3.25.0
colorama==0.4.6
ConfigSpace==0.6.1
contextlib2==21.6.0
contourpy==1.0.7
cycler==0.11.0
dask==2023.5.0
dask-jobqueue==0.8.1
datasets==2.9.0
dill==0.3.6
distributed==2023.5.0
docker-pycreds==0.4.0
emcee==3.1.4
filelock==3.9.0
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.1.0
gitdb==4.0.10
GitPython==3.1.30
huggingface-hub==0.12.0
idna==3.4
importlib-metadata==6.6.0
Jinja2==3.1.2
joblib==1.2.0
json-tricks==3.16.1
kiwisolver==1.4.4
lit==16.0.5
locket==1.0.0
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib==3.7.1
more-itertools==9.1.0
mpmath==1.2.1
msgpack==1.0.5
multidict==6.0.4
multiprocess==0.70.14
networkx==3.0rc1
nni==2.10
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==23.1
pandas==2.0.2
partd==1.4.0
pathtools==0.1.2
Pillow==9.5.0
protobuf==4.21.12
psutil==5.9.4
pyarrow==11.0.0
pycodestyle==2.10.0
pycryptodomex==3.17
pynisher==1.0.5
pyparsing==3.0.9
pyrfr==0.9.0
python-dateutil==2.8.2
PythonWebHDFS==0.2.3
pytorch-triton==2.0.0+0d7e753227
pytz==2023.3
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
responses==0.18.0
safetensors==0.3.1
schema==0.7.5
scikit-learn==1.2.1
scipy==1.10.1
sentry-sdk==1.15.0
setproctitle==1.3.2
six==1.16.0
-e git+https://github.com/automl/SMAC3.git@d5d54562333c6845c06227de15faec30ef553d23#egg=smac
smmap==5.0.0
sortedcontainers==2.4.0
sympy==1.11.1
tblib==1.7.0
threadpoolctl==3.1.0
tiktoken==0.2.0
timm==0.9.2
tokenizers==0.13.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
torchsummary==1.5.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.64.1
transformers==4.26.0
triton==2.0.0
typeguard==2.13.3
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.14
wandb==0.13.10
websockets==10.4
xxhash==3.2.0
yarl==1.8.2
zict==3.0.0
zipp==3.15.0
